feat: take out filters out to component
