+++ src/pages/PetsList/PetsList.tsx
@@ -1,44 +1,21 @@
 import { FiFilter } from 'react-icons/fi';
 import { CustomButton } from 'components/CustomButton';
-import {
-  useGetFilteredAnimalsQuery,
-  type SortOrder,
-} from 'src/redux/animals/animalsApi';
+import { useGetFilteredAnimalsQuery } from 'src/redux/animals/animalsApi';
 import AnimalCard from 'components/AnimalCard';
 import { PetsListSkeleton } from 'components/sceletons/PetsListSkeleton';
 import Pagination from 'components/Pagination';
-import { useCallback, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useNavigate, useSearchParams } from 'react-router';
 import { showToast } from 'components/Toast';
-import FilterItem from 'components/FilterItem';
-import {
-  ageOption,
-  animalTypeOptions,
-  AnimalType,
-  genderOption,
-  size,
-} from '../Announcement/types';
-import BreedSelect from 'components/BreedSelect';
-import { CitySelect } from 'components/CitySelect';
-import { Controller, useForm } from 'react-hook-form';
 import { cn } from 'components/lib/utils';
 import { useWindowSize } from '@uidotdev/usehooks';
-import FileterLabel from 'components/FileterLabel';
-import { getActiveFilters, mapAnimalType } from './mapping';
+import { mapAnimalType } from './mapping';
 import { Spinner } from 'components/Spinner';
 import CloseSVG from 'src/assets/CloseSVG';
+import Filter, { type FilterFormValues } from 'components/Filter';
+import { useFilters } from 'src/context/FiltersContext';
 
 const limit = 12;
-
-interface FilterFormValues {
-  animalType: AnimalType | undefined;
-  gender: string;
-  breed: string;
-  location: string;
-  age: string;
-  size: string;
-  sortByDate?: 'newest' | 'oldest';
-}
 
 const PetsList = () => {
   const [searchParams, setSearchParams] = useSearchParams();
@@ -46,32 +23,17 @@ const PetsList = () => {
   const [page, setPage] = useState(rawPage === 0 ? 1 : rawPage);
   const [openFilters, setOpenFilters] = useState(false);
   const [openSorting, setOpenSorting] = useState(false);
-  const [filtersParams, setFiltersParams] = useState<Partial<FilterFormValues>>(
-    {}
-  );
   const [isFilterApplied, setIsFilterApplied] = useState(false);
-  const [sorting, setSorting] = useState<SortOrder>('newest');
+  const { filtersParams, setFiltersParams } = useFilters();
 
   const navigate = useNavigate();
   const windowSize = useWindowSize();
 
   const { data, isLoading, isFetching, error } = useGetFilteredAnimalsQuery(
-    { page, limit, ...filtersParams, sortByDate: sorting },
+    { page, limit, ...filtersParams },
     { skip: !filtersParams }
   );
 
-  const { control, handleSubmit, watch, reset, resetField } =
-    useForm<FilterFormValues>({
-      defaultValues: {
-        animalType: undefined,
-        gender: '',
-        breed: '',
-        location: '',
-        age: '',
-        size: '',
-      },
-    });
-  const selectedAnimalType = watch('animalType');
   const totalPages = data && Math.ceil(data?.total / limit);
   const title =
     filtersParams && filtersParams?.animalType
@@ -119,38 +81,18 @@ const PetsList = () => {
   };
 
   const handleAscSorting = () => {
-    setSorting('newest');
-    setFiltersParams({ sortByDate: 'newest' });
+    setFiltersParams({ ...filtersParams, sortByDate: 'newest' });
     setOpenSorting(false);
   };
   const handleDescSorting = () => {
-    setSorting('oldest');
-    setFiltersParams({ sortByDate: 'oldest' });
+    setFiltersParams({ ...filtersParams, sortByDate: 'oldest' });
     setOpenSorting(false);
   };
-  const handleClearFilter = () => {
-    setFiltersParams({});
-    reset();
-  };
-  const handleDeleteFilterItem = useCallback(
-    (item: keyof FilterFormValues) => {
-      setFiltersParams(prev => {
-        const newParams = { ...prev };
-        delete newParams[item];
-        return newParams;
-      });
-      resetField(item);
-      setOpenFilters(false);
-    },
-    [resetField]
-  );
-  const activeFilterItems = getActiveFilters(filtersParams);
-
   return (
     <div className="container">
       <div className="relative flex justify-center mt-30 lg:mt-100 mb-100 lg:mb-50">
         <div className="flex flex-col">
-          <h1 className="mb-10 md:text-[32px] text-lg text-default-btn dark:text-default-btn">
+          <h1 className="mb-10 text-default-btn md:text-[32px] dark:text-default-btn text-lg">
             {title}
           </h1>
           {isFilterApplied && data && (
@@ -161,17 +103,7 @@ const PetsList = () => {
             </p>
           )}
         </div>
-        {data?.total === 0 ? (
-          <></>
-        ) : (
-          // <CustomButton
-          //   type="button"
-          //   styleType="defaultButton"
-          //   className="top-[95px] md:top-[135px] left-[calc(50%-98px)] z-50 absolute m-0 w-[196px]"
-          //   onClick={handleClearFilter}
-          // >
-          //   До списку тварин
-          // </CustomButton>
+        {data && data?.total > 0 && (
           <div
             className={`lg:top-0 absolute flex justify-between m-0 w-full ${isFilterApplied ? 'top-80' : 'top-50'}`}
           >
@@ -188,7 +120,7 @@ const PetsList = () => {
               <CustomButton
                 type="button"
                 styleType="whiteButton"
-                className="w-[192px] md:w-[217px]  text-medium text-base text-default-btn"
+                className="w-[192px] md:w-[217px] text-default-btn text-medium text-base"
                 onClick={() => setOpenSorting(prev => !prev)}
               >
                 Сортування за датою
@@ -199,7 +131,7 @@ const PetsList = () => {
                     onClick={handleAscSorting}
                     className={cn(
                       'text-default-btn text-left text-lg focus:outline-none hover:text-orange transition-all duration-300',
-                      sorting === 'newest' && 'text-orange'
+                      filtersParams?.sortByDate === 'newest' && 'text-orange'
                     )}
                   >
                     Останні оголошення
@@ -208,7 +140,7 @@ const PetsList = () => {
                     onClick={handleDescSorting}
                     className={cn(
                       'text-default-btn text-left text-lg focus:outline-none hover:text-orange transition-all duration-300',
-                      sorting === 'oldest' && 'text-orange'
+                      filtersParams?.sortByDate === 'oldest' && 'text-orange'
                     )}
                   >
                     Давні оголошення
@@ -232,8 +164,8 @@ const PetsList = () => {
               }}
             >
               <div className="relative flex flex-col bg-dialog xl:bg-transparent xl:p-0 pt-32 rounded-4xl">
-                <div className="flex items-center justify-between px-16 mb-4">
-                  <div className="block xl:hidden text-default-btn text-base font-medium">
+                <div className="flex justify-between items-center mb-4 px-16">
+                  <div className="xl:hidden block font-medium text-default-btn text-base">
                     Фільтр
                   </div>
                   <div
@@ -243,120 +175,13 @@ const PetsList = () => {
                     <CloseSVG fill="white" size="27" />
                   </div>
                 </div>
-                <div className="flex flex-wrap gap-x-16 gap-y-10 xl:mb-32 ml-16 xl:ml-0 max-w-[344px]">
-                  {activeFilterItems.length > 0 &&
-                    activeFilterItems.map(({ key, value }) => (
-                      <FileterLabel
-                        key={key}
-                        label={value}
-                        onRemove={(e: React.MouseEvent<HTMLButtonElement>) => {
-                          e.stopPropagation();
-                          handleDeleteFilterItem(key as keyof FilterFormValues);
-                        }}
-                      />
-                    ))}
-                </div>
-                <form
-                  onSubmit={handleSubmit(onSubmit)}
-                  onClick={e => e.stopPropagation()}
-                  className={cn(
-                    'xl:flex flex-col transition-all duration-500 xl:bg-transparent',
-                    'xl:static  xl:gap-32',
-                    'flex flex-col bg-dialog p-16 gap-16 rounded-4xl w-[344px] xl:w-[306px]  xl:mt-0  xl:p-0'
-                  )}
-                >
-                  <Controller
-                    name="animalType"
-                    control={control}
-                    render={({ field }) => (
-                      <FilterItem
-                        {...field}
-                        label="Вид тварини"
-                        items={animalTypeOptions}
-                        className="text-default-btn"
-                      />
-                    )}
-                  />
-                  <Controller
-                    name="gender"
-                    control={control}
-                    render={({ field }) => (
-                      <FilterItem
-                        {...field}
-                        label="Стать"
-                        items={genderOption}
-                        className="text-default-btn"
-                      />
-                    )}
-                  />
-                  <Controller
-                    name="breed"
-                    control={control}
-                    render={({ field }) => (
-                      <BreedSelect
-                        {...field}
-                        className="w-full h-[40px]"
-                        type={selectedAnimalType}
-                        placeholder="Порода"
-                      />
-                    )}
-                  />
-                  <Controller
-                    name="location"
-                    control={control}
-                    render={({ field }) => (
-                      <CitySelect
-                        {...field}
-                        className="h-[40px]"
-                        widthClass="w-full"
-                        placeholder="Місто"
-                      />
-                    )}
-                  />
-                  <Controller
-                    name="age"
-                    control={control}
-                    render={({ field }) => (
-                      <FilterItem
-                        {...field}
-                        label="Вік"
-                        items={ageOption}
-                        className="text-default-btn"
-                      />
-                    )}
-                  />
 
-                  <Controller
-                    name="size"
-                    control={control}
-                    render={({ field }) => (
-                      <FilterItem
-                        {...field}
-                        label="Розмір"
-                        items={size}
-                        className="text-default-btn"
-                      />
-                    )}
-                  />
-
-                  <CustomButton
-                    type="submit"
-                    styleType="defaultButton"
-                    className="self-center m-0"
-                    loading={isLoading || isFetching}
-                  >
-                    Застосувати фільтр
-                  </CustomButton>
-                  <CustomButton
-                    type="button"
-                    styleType="whiteButton"
-                    className="self-center m-0 xl:-mt-12 mb-50 lg:mb-100"
-                    onClick={handleClearFilter}
-                    disabled={Object.keys(filtersParams).length === 0}
-                  >
-                    Очистити фільтр
-                  </CustomButton>
-                </form>
+                <Filter
+                  onSubmit={onSubmit}
+                  isLoading={isLoading}
+                  isFetching={isFetching}
+                  onClose={() => setOpenFilters(false)}
+                />
               </div>
             </div>
           )}
