+++ src/pages/PetsList/PetsPage.tsx
@@ -1,0 +1,12 @@
+import { FiltersProvider } from 'src/context/FiltersContext';
+import PetsList from './PetsList';
+
+const PetsPage = () => {
+  return (
+    <FiltersProvider>
+      <PetsList />
+    </FiltersProvider>
+  );
+};
+
+export default PetsPage;
